implemented regions.py for charts and data
